pyproject.toml
[project]
name = "codee_test"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "HanRix1", email = "HanRix1@mail.ru"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.2",
    "pydantic-settings>=2.5.2",
    "psycopg2>=2.9.9",
    "psycopg-binary>=3.2.2",
    "asyncpg>=0.29.0",
    "aiohttp>=3.10.5",
    "requests>=2.32.3",
    "PyJWT>=2.9.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.5",
    "pytest>=8.3.3",
    "mypy>=1.11.2",
    "isort>=5.13.2",
]
